npus-ai-aiops-ibm-poc-main/tools-call-sample-app/pyproject.toml
[project]
name = "agent-eval"
version = "0.1.0"
description = "AgentTool Eval App - Reference implementation for AI agents with tool-calling evaluation (LangGraph, MCP, Langfuse, RAGAS, DeepEval)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-mcp-adapters>=0.1.0",
    "langfuse>=2.0.0,<3.0.0",
    "ragas>=0.2.0",
    "deepeval>=2.0.0",
    "mcp>=1.0.0",
    "streamlit>=1.40.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.30.0",
    "nest-asyncio>=1.6.0",
    "databricks-sdk>=0.20.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent_eval", "evaluation"]
